opcode使用枚举类型

+++ Unity/Assets/Scripts/Handler/Actor_CreateUnitsHandler.cs
@@ -2,7 +2,7 @@
 
 namespace Model
 {
-	[MessageHandler(Opcode.Actor_CreateUnits)]
+	[MessageHandler((int)Opcode.Actor_CreateUnits)]
 	public class Actor_CreateUnitsHandler : AMHandler<Actor_CreateUnits>
 	{
 		protected override void Run(Actor_CreateUnits message)
